05_cloud_pipelines/src/requirements.txt
beautifulsoup4==4.*
Flask==2.*
functions_framework==3.*
geopy==2.*
lat_lon_parser==1.*
pandas==2.*
python-dotenv==1.*
pytz
Requests==2.*
SQLAlchemy==1.*
timezonefinder==6.*
pymysql==1.*
google-cloud-storage==2.*
nbconvert==6.*
nbformat==5.*
ipykernel==6.*
ipython==8.*
numpy==1.*
matplotlib==3.*
seaborn==0.*
